# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.ErrorBody'
      meta: {}
      success:
        type: boolean
    type: object
  github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.ErrorBody:
    properties:
      code:
        type: string
      details:
        description: optional field-level errors
      message:
        type: string
    type: object
...
  internal_admin_controller.subUsageSummary:
    properties:
      id:
        type: string
      is_active:
        type: boolean
      max_messages:
        type: integer
      messages_used:
        type: integer
      tier:
        type: string
    type: object
  internal_admin_controller.userUsageResponse:
    properties:
      free_messages_used:
        type: integer
      subscriptions:
        items:
          $ref: '#/definitions/internal_admin_controller.subUsageSummary'
        type: array
      total_messages:
        type: integer
      user_id:
        type: string
    type: object
  internal_auth_controller.setRoleRequest:
    properties:
      role:
        description: Role must be "user" or "admin".
        enum:
        - user
        - admin
        example: admin
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Production-grade AI chat backend with Auth0 authentication, DDD, quota
    management, and subscription bundles.
  title: Secure AI Chat Backend
  version: "1.0"
paths:
...
  /api/v1/admin/users:
    get:
      description: Returns all registered users ordered by creation date. Admin role
        required.
      parameters:
      - description: Unix timestamp
        in: header
        name: X-Request-Timestamp
        required: true
        type: string
      - description: Unique nonce
        in: header
        name: X-Nonce
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "403":
          description: Forbidden — admin role required
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - admin
  /api/v1/admin/users/{id}/role:
    patch:
      consumes:
      - application/json
      description: Promotes or demotes a user to/from admin. Role must be "user" or
        "admin". Admin role required.
      parameters:
      - description: Unix timestamp (seconds since epoch, e.g. 1711234567)
        in: header
        name: X-Request-Timestamp
        required: true
        type: string
      - description: Unique string per request (e.g. uuid or random hex)
        in: header
        name: X-Nonce
        required: true
        type: string
      - description: Target user UUID
        in: path
        name: id
        required: true
        type: string
      - description: Role assignment payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth_controller.setRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "400":
          description: Invalid role or user ID
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "403":
          description: Forbidden — admin role required
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: Set user role
      tags:
      - admin
  /api/v1/admin/users/{id}/usage:
    get:
      description: 'Returns detailed usage stats for a specific user: free quota used
        this month, total messages, and all subscription bundles. Admin role required.'
      parameters:
      - description: Unix timestamp
        in: header
        name: X-Request-Timestamp
        required: true
        type: string
      - description: Unique nonce
        in: header
        name: X-Nonce
        required: true
        type: string
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User usage details
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_admin_controller.userUsageResponse'
              type: object
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "403":
          description: Forbidden — admin role required
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get user usage
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: Auth0 JWT — prefix with "Bearer ". Obtain via Auth0 dashboard (APIs
      → Test tab) or your Auth0 login flow.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"